Remove dependency on happy >= 1.19
The hackage package contains the generated parser, so it doesn't need any
happy at all.

To build from source, one does need happy, but it does not matter which
version.  However, the older happy generated code that does not work with
GHC 7.8, so if you made a hackage package on a machine with an older happy,
you get a package that does not work with newer GHC.


Sigh.